wow even more updates!

+++ main.py
@@ -1,24 +1,70 @@
+import json
+from typing import Union
 from fastapi import FastAPI
 import db
+from pydantic import BaseModel
+from sqlite3 import IntegrityError
+
+
+class Secret(BaseModel):
+    organization: str = "default"
+    username: Union[str, None]
+    password: Union[str, None]
+
+    def __iter__(self):
+        return iter((self.organization, self.username, self.password))
+
+
+class Message(BaseModel):
+    message: str
+
 
 app = FastAPI()
 
-@app.get("/add/{organization}/{username}/{password}")
-async def add_password(organization: str, username: str, password: str):
-    return db.add_password(username=username, password=password, organization=organization)
 
-@app.get("/delete/{organization}/{username}")
-async def delete_password(organization: str, username: str):
-    return db.delete_password(username=username, organization=organization)
+@app.post("/add/", response_model=Message)
+async def add_password(secret: Secret):
+    organization, username, password = secret
 
-@app.get("/update/{organization}/{username}/{password}")
-async def update_password(organization: str, username: str, password: str):
-    return db.update_password(username=username, password=password, organization=organization)
+    if password is None:
+        return Message(message="Password required")
+    if username is None:
+        return Message(message="Password required")
+
+    try:
+        db.add_password(username=username, password=password, organization=organization)
+    except IntegrityError:
+        return Message(message="Password already exists.")
+
+    return Message(message=f"Password added to {organization} for username {username}")
+
+
+@app.delete("/delete/")
+async def delete_password(secret: Secret):
+    organization, username, _ = secret
+
+    if username is None:
+        return Message(message="Username required")
+
+    db.delete_password(username=username, organization=organization)
+
+    return Message(message=f"Username deleted from {organization}")
+
+
+@app.put("/update/")
+async def update_password(secret: Secret):
+    organization, username, password = secret
+
+    db.update_password(username=username, password=password, organization=organization)
+
+    return Message(message=f"Password updated for {username} in {organization}")
+
 
 @app.get("/get/{organization}/{username}")
-async def get_password(organization: str, username: str):
+async def get_password(username: str, organization: str = "default"):
     return db.get_password(username=username, organization=organization)
 
-@app.get("/get/{organization}")
-async def get_organization_usernames(organization: str):
-    return db.get_organization_usernames(organization=organization)
+
+# @app.get("/get/{organization}")
+# async def get_organization_usernames(organization: str):
+#     return db.get_organization_usernames(organization=organization)
